Update de.json (#44)
This commit updates several German translations to make them more
consistent and better fit the available space.

Some strings are changed from singular to plural as they address
multiple items. English is neutral here without an article which does
not work in German.

The abbreviations for "Zunehmender" and "Abnehmender" from the previous
commit are reversed as there is enough space for them to fit.