Fix machine translations for proper German grammar (#53)
The usual output of Google Translate does not match the grammar and
style that is common in German UIs. I also flips the perspective in
several translations asking the user to do something which the computer
is about to perform.

In addition I prefer to translate "show" / "hide" as "einblenden" /
"ausblenden" (Mac style) where they are the two options for showing
something on screen. "Anzeigen" which machine translations prefer is
more for items that are opened in a separate view etc.

Finally the machine translations favor the formal "Sie" while Home
Assistant uses the less formal "Du" ("you") instead.